Update ReplaceTest
Fixing little issues with the serialized data replacing.
More info:

The logic however is right, the recursive function try to unpack
the data until it's get a simple string value, make the replace and return up.

The issue is in the order we pass the `$serialised` parameter.
Now every recursive call to `recursive_unserialize_replace`
use always `false`, `true` must be set when we call the function outside,
so we can retrieve the serialized value "correctly replaced" (tests permits)
internally we try to unserialize the data the first time,
then when every call return up to the caller we have the first `$serialised`
set to `true` and the data will be serialized accordingly.

I've only some doubts about the use of `is_serialized`, `is_serialized_string` and the use of `unserialize`, `serialize`.

There was a comment within `recursive_unserialize_replace` that tell to switch to
`maybe_unserialize` because serialized data by WordPress is different
than the serialized data by `serialize` but I'm not sure,
WordPress internally use `serialize` (in wrong way I think because permit
object serialization and deserialization aka object injection but since we should work with it.

Also, right know I put the same code used in WordPress for:
 - `is_serialized_string`,
 - `is_serialized`
 - `maybe_serialize`
 because any other stub will not well defined as WordPress do and since
 serializing data is a sensitive data I think in this case better to use
 the original functions as if it were native functions.

 Also, removed `add_error` undefined method call and replaced with
 a throw for when `WP_DEBUG` is active and an action when not.
 See catch block.